+++ app/Http/Controllers/Purchase/PurchaseDownPaymentController.php
@@ -600,7 +600,7 @@ class PurchaseDownPaymentController extends Controller
         $query = PurchaseDownPayment::where('code',CustomHelper::decrypt($request->id))->first();
         
         if($query) {
-            if($query->status == '5'){
+            if(in_array($query->status,['4','5'])){
                 $response = [
                     'status'  => 500,
                     'message' => 'Data telah ditutup anda tidak bisa menutup lagi.'
@@ -642,7 +642,7 @@ class PurchaseDownPaymentController extends Controller
     public function destroy(Request $request){
         $query = PurchaseDownPayment::where('code',CustomHelper::decrypt($request->id))->first();
 
-        if($query->approval() || in_array($query->status,['2','3'])){
+        if($query->approval()){
             foreach($query->approval()->approvalMatrix as $row){
                 if($row->status == '2'){
                     return response()->json([
@@ -651,6 +651,13 @@ class PurchaseDownPaymentController extends Controller
                     ]);
                 }
             }
+        }
+
+        if(in_array($query->status,['2','3'])){
+            return response()->json([
+                'status'  => 500,
+                'message' => 'Jurnal sudah dalam progres, anda tidak bisa melakukan perubahan.'
+            ]);
         }
         
         if($query->delete()) {
